moved raising error per tests
