+++ lib/capybara_screenshot_diff/rspec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "minitest"
+
+module Capybara::Screenshot::Diff
+  ASSERTION = ::StandardError
+end
+
+require_relative "minitest"
